requirements.txt
Flask==2.0.2
Flask-SQLAlchemy==2.3.2
requests==2.20.1
SQLAlchemy==1.2.14
urllib3==1.24.1
PyYAML==3.13